% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stats-v15.9.10.R
\name{get.normalised.dari.templates}
\alias{get.normalised.dari.templates}
\title{Esta función normaliza los templates del ARI propuesto por Asslid}
\usage{
get.normalised.dari.templates(time.instants, abp.signal, sample.release,
  sampling.time, baseline.initial.time = time.instants[1],
  baseline.final.time = time.instants[sample.release],
  min.abp.max.delta.time = 5 * 0.8, min.cbfv.max.delta.time = 8 * 0.8,
  stabilisation.time = 1, at.param.rounding.digits = 6,
  normalised.cbfv.rounding.digits = 4, time.tol = sampling.time/100, ...)
}
\arguments{
\item{time.tol}{}
}
\value{
Los templates de ARI normalizados
}
\description{
Esta función normaliza los templates del ARI propuesto por Asslid
}

